See jkroes/dotfiles/init.vim
This change is meant to integrate with my nvim config file, which sets
the value of the backup variable on startup so that it can be reset by
the ranger wrapper after each ranger invocation. The init file includes
bindings that set the value of the floaterm-provided variable before
opening files in ranger, in order to open files in splits and new tabs.